Fix cast issue to allow for actual event codes to be read
No idea why this wasn't working previously, but debug logging with the
old approach always printed the values for the HCI event type and
payload, which was clearly wrong, since other consumers were reacting to
the message in different ways.

Copying the cast from the serial_service fixed this. yayyy. Love to
spend a week and a half of time on an issue that wasn't actually an
issue.